Removed fetching from TMCoreData Object
the fetching code is now part of the NSManagedObjectContext category,
this way you can signal which context you wish to perform the
perform/save/complete on.

NOTE: the complete is called within the thread context of the
managedobjectcontext, this might be the main thread, or a background
thread.

The complete routine is always called AFTER the save has completed
successfully